moved VisGeometry
